Copy sanitize_callable_params to new test file
